+++ packages/google_maps_flutter/google_maps_flutter_android/example/lib/scrolling_map.dart
@@ -25,7 +25,7 @@ class ScrollingMapPage extends GoogleMapExampleAppPage {
 }
 
 class ScrollingMapBody extends StatelessWidget {
-  const ScrollingMapBody({Key? key}) : super(key: key);
+  const ScrollingMapBody({super.key});
 
   @override
   Widget build(BuildContext context) {