+++ packages/video_player/video_player_android/example/lib/mini_controller.dart
@@ -219,6 +219,7 @@ class MiniController extends ValueNotifier<VideoPlayerValue> {
     final Completer<void> initializingCompleter = Completer<void>();
 
     void eventListener(VideoEvent event) {
+      // ignore: missing_enum_constant_in_switch
       switch (event.eventType) {
         case VideoEventType.initialized:
           value = value.copyWith(
@@ -348,7 +349,7 @@ class MiniController extends ValueNotifier<VideoPlayerValue> {
 /// Widget that displays the video controlled by [controller].
 class VideoPlayer extends StatefulWidget {
   /// Uses the given [controller] for all video rendered in this widget.
-  const VideoPlayer(this.controller, {Key? key}) : super(key: key);
+  const VideoPlayer(this.controller, {super.key});
 
   /// The [MiniController] responsible for the video being rendered in
   /// this widget.
@@ -447,7 +448,7 @@ class _VideoScrubberState extends State<_VideoScrubber> {
 class VideoProgressIndicator extends StatefulWidget {
   /// Construct an instance that displays the play/buffering status of the video
   /// controlled by [controller].
-  const VideoProgressIndicator(this.controller, {Key? key}) : super(key: key);
+  const VideoProgressIndicator(this.controller, {super.key});
 
   /// The [MiniController] that actually associates a video with this
   /// widget.