+++ packages/webview_flutter/webview_flutter_android/lib/src/legacy/webview_android_widget.dart
@@ -23,10 +23,12 @@ class WebViewAndroidWidget extends StatefulWidget {
     required this.callbacksHandler,
     required this.javascriptChannelRegistry,
     required this.onBuildWidget,
-    @visibleForTesting this.webViewProxy = const WebViewProxy(),
+    @visibleForTesting
+        this.webViewProxy = const WebViewProxy(),
     @visibleForTesting
         this.flutterAssetManager = const android_webview.FlutterAssetManager(),
-    @visibleForTesting this.webStorage,
+    @visibleForTesting
+        this.webStorage,
   });
 
   /// Initial parameters used to setup the WebView.
@@ -101,10 +103,12 @@ class WebViewAndroidPlatformController extends WebViewPlatformController {
     required CreationParams creationParams,
     required this.callbacksHandler,
     required this.javascriptChannelRegistry,
-    @visibleForTesting this.webViewProxy = const WebViewProxy(),
+    @visibleForTesting
+        this.webViewProxy = const WebViewProxy(),
     @visibleForTesting
         this.flutterAssetManager = const android_webview.FlutterAssetManager(),
-    @visibleForTesting android_webview.WebStorage? webStorage,
+    @visibleForTesting
+        android_webview.WebStorage? webStorage,
   })  : webStorage = webStorage ?? android_webview.WebStorage.instance,
         assert(creationParams.webSettings?.hasNavigationDelegate != null),
         super(callbacksHandler) {
@@ -618,15 +622,13 @@ class WebViewProxy {
       android_webview.WebView webView,
       android_webview.WebResourceRequest request,
       android_webview.WebResourceError error,
-    )?
-        onReceivedRequestError,
+    )? onReceivedRequestError,
     void Function(
       android_webview.WebView webView,
       int errorCode,
       String description,
       String failingUrl,
-    )?
-        onReceivedError,
+    )? onReceivedError,
     void Function(android_webview.WebView webView,
             android_webview.WebResourceRequest request)?
         requestLoading,