+++ packages/google_sign_in/google_sign_in/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Aligns Dart and Flutter SDK constraints.
+
 ## 6.0.1
 
 * Updates links for the merge of flutter/plugins into flutter/packages.