+++ packages/google_sign_in/google_sign_in/CHANGELOG.md
@@ -1,3 +1,15 @@
+## NEXT
+
+* Aligns Dart and Flutter SDK constraints.
+
+## 6.0.2
+
+* Updates iOS minimum version in README.
+
+## 6.0.1
+
+* Updates links for the merge of flutter/plugins into flutter/packages.
+
 ## 6.0.0
 
 * **Breaking change** for platform `web`: